Refactor issue #638 state action exports
