+++ src/zivo/state/__init__.py
@@ -1,159 +1,5 @@
 """State management primitives."""
 
-from .actions import (
-    Action,
-    ActivateNextTab,
-    ActivatePreviousTab,
-    AddBookmark,
-    ArchiveExtractCompleted,
-    ArchiveExtractFailed,
-    ArchiveExtractProgress,
-    ArchivePreparationCompleted,
-    ArchivePreparationFailed,
-    BeginBookmarkSearch,
-    BeginCommandPalette,
-    BeginCreateInput,
-    BeginDeleteTargets,
-    BeginEmptyTrash,
-    BeginExtractArchiveInput,
-    BeginFileSearch,
-    BeginFilterInput,
-    BeginFindAndReplace,
-    BeginGoToPath,
-    BeginGrepReplace,
-    BeginGrepReplaceSelected,
-    BeginGrepSearch,
-    BeginHistorySearch,
-    BeginRenameInput,
-    BeginSelectedFilesGrep,
-    BeginShellCommandInput,
-    BeginTextReplace,
-    BeginZipCompressInput,
-    BrowserSnapshotFailed,
-    BrowserSnapshotLoaded,
-    CancelArchiveExtractConfirmation,
-    CancelCommandPalette,
-    CancelDeleteConfirmation,
-    CancelEmptyTrashConfirmation,
-    CancelFilterInput,
-    CancelPasteConflict,
-    CancelPendingInput,
-    CancelShellCommandInput,
-    CancelZipCompressConfirmation,
-    ChildPaneSnapshotFailed,
-    ChildPaneSnapshotLoaded,
-    ClearPendingKeySequence,
-    ClearSelection,
-    ClipboardPasteCompleted,
-    ClipboardPasteFailed,
-    ClipboardPasteNeedsResolution,
-    CloseCurrentTab,
-    ConfigSaveCompleted,
-    ConfigSaveFailed,
-    ConfirmArchiveExtract,
-    ConfirmDeleteTargets,
-    ConfirmEmptyTrash,
-    ConfirmFilterInput,
-    ConfirmZipCompress,
-    CopyPathsToClipboard,
-    CopyTargets,
-    CutTargets,
-    CycleConfigEditorValue,
-    CycleFindReplaceField,
-    CycleGrepReplaceField,
-    CycleGrepReplaceSelectedField,
-    CycleGrepSearchField,
-    CycleReplaceField,
-    CycleSelectedFilesGrepField,
-    DeletePendingInputForward,
-    DirectorySizesFailed,
-    DirectorySizesLoaded,
-    DismissAttributeDialog,
-    DismissConfigEditor,
-    DismissNameConflict,
-    EnterCursorDirectory,
-    ExitCurrentPath,
-    ExternalLaunchCompleted,
-    ExternalLaunchFailed,
-    FileMutationCompleted,
-    FileMutationFailed,
-    FileSearchCompleted,
-    FileSearchFailed,
-    FocusSplitTerminal,
-    GoBack,
-    GoForward,
-    GoToHomeDirectory,
-    GoToParentDirectory,
-    GrepSearchCompleted,
-    GrepSearchFailed,
-    InitializeState,
-    JumpCursor,
-    MoveCommandPaletteCursor,
-    MoveConfigEditorCursor,
-    MoveCursor,
-    MoveCursorAndSelectRange,
-    MoveCursorByPage,
-    MovePendingInputCursor,
-    OpenFindResultInEditor,
-    OpenGrepResultInEditor,
-    OpenNewTab,
-    OpenPathInEditor,
-    OpenPathWithDefaultApp,
-    OpenTerminalAtPath,
-    PasteClipboard,
-    PasteIntoPendingInput,
-    ReloadDirectory,
-    RemoveBookmark,
-    RequestBrowserSnapshot,
-    RequestDirectorySizes,
-    ResolvePasteConflict,
-    SaveConfigEditor,
-    SelectAllVisibleEntries,
-    SelectedFilesGrepKeywordChanged,
-    SendSplitTerminalInput,
-    SetCommandPaletteQuery,
-    SetCursorPath,
-    SetFilterQuery,
-    SetFindReplaceField,
-    SetGrepReplaceField,
-    SetGrepReplaceSelectedField,
-    SetGrepSearchField,
-    SetNotification,
-    SetPendingInputCursor,
-    SetPendingInputValue,
-    SetPendingKeySequence,
-    SetReplaceField,
-    SetShellCommandValue,
-    SetSort,
-    SetTerminalHeight,
-    SetUiMode,
-    ShellCommandCompleted,
-    ShellCommandFailed,
-    ShowAttributes,
-    SplitTerminalExited,
-    SplitTerminalOutputReceived,
-    SplitTerminalStarted,
-    SplitTerminalStartFailed,
-    SubmitCommandPalette,
-    SubmitPendingInput,
-    SubmitShellCommand,
-    TextReplaceApplied,
-    TextReplaceApplyFailed,
-    TextReplacePreviewCompleted,
-    TextReplacePreviewFailed,
-    ToggleHiddenFiles,
-    ToggleSelection,
-    ToggleSelectionAndAdvance,
-    ToggleSplitTerminal,
-    UndoCompleted,
-    UndoFailed,
-    UndoLastOperation,
-    ZipCompressCompleted,
-    ZipCompressFailed,
-    ZipCompressPreparationCompleted,
-    ZipCompressPreparationFailed,
-    ZipCompressProgress,
-)
 from .effects import (
     CloseSplitTerminalEffect,
     Effect,
@@ -201,6 +47,7 @@ from .models import (
     FilterState,
     FindReplaceFieldId,
     GrepReplaceFieldId,
+    GrepReplaceSelectedFieldId,
     GrepSearchFieldId,
     GrepSearchResultState,
     HistoryState,
@@ -249,237 +96,84 @@ from .selectors import (
 )
 
 __all__ = [
-    "Action",
-    "ActivateNextTab",
-    "ActivatePreviousTab",
     "AppState",
-    "ArchiveExtractCompleted",
     "ArchiveExtractConfirmationState",
-    "ArchiveExtractFailed",
-    "ArchiveExtractProgress",
     "ArchiveExtractProgressState",
-    "ArchivePreparationCompleted",
-    "ArchivePreparationFailed",
     "AttributeInspectionState",
-    "BeginCommandPalette",
-    "BeginFilterInput",
-    "BeginCreateInput",
-    "BeginDeleteTargets",
-    "BeginEmptyTrash",
-    "BeginBookmarkSearch",
-    "BeginExtractArchiveInput",
-    "BeginZipCompressInput",
-    "BeginFileSearch",
-    "BeginGoToPath",
-    "BeginGrepSearch",
-    "BeginHistorySearch",
-    "BeginRenameInput",
-    "BeginTextReplace",
-    "BeginFindAndReplace",
-    "BeginGrepReplace",
-    "BeginGrepReplaceSelected",
-    "BeginSelectedFilesGrep",
-    "CycleSelectedFilesGrepField",
-    "SelectedFilesGrepKeywordChanged",
-    "BeginShellCommandInput",
-    "BrowserTabState",
     "BrowserSnapshot",
-    "BrowserSnapshotFailed",
-    "BrowserSnapshotLoaded",
-    "CancelArchiveExtractConfirmation",
-    "CancelZipCompressConfirmation",
-    "CancelCommandPalette",
-    "CancelDeleteConfirmation",
-    "CancelEmptyTrashConfirmation",
-    "CancelPasteConflict",
-    "CancelFilterInput",
-    "ClearPendingKeySequence",
-    "CancelPendingInput",
-    "CancelShellCommandInput",
-    "ChildPaneSnapshotFailed",
-    "ChildPaneSnapshotLoaded",
-    "CloseCurrentTab",
-    "ClearSelection",
+    "BrowserTabState",
     "ClipboardState",
+    "CloseSplitTerminalEffect",
     "CommandPaletteSource",
     "CommandPaletteState",
-    "ClipboardPasteCompleted",
-    "ClipboardPasteFailed",
-    "ClipboardPasteNeedsResolution",
     "ConfigEditorState",
     "CurrentPaneDeltaState",
-    "ConfigSaveCompleted",
-    "ConfigSaveFailed",
-    "ConfirmArchiveExtract",
-    "ConfirmZipCompress",
-    "ConfirmDeleteTargets",
-    "ConfirmEmptyTrash",
-    "ConfirmFilterInput",
-    "CopyPathsToClipboard",
-    "CopyTargets",
-    "CycleGrepSearchField",
-    "CycleGrepReplaceField",
-    "CycleGrepReplaceSelectedField",
-    "CycleFindReplaceField",
-    "CycleReplaceField",
-    "CutTargets",
-    "DismissAttributeDialog",
-    "DismissConfigEditor",
     "DeleteConfirmationState",
-    "DeletePendingInputForward",
     "DirectoryEntryState",
-    "EmptyTrashConfirmationState",
     "DirectorySizeCacheEntry",
     "DirectorySizeDeltaState",
     "DirectorySizeStatus",
-    "DirectorySizesFailed",
-    "DirectorySizesLoaded",
-    "DismissNameConflict",
-    "EnterCursorDirectory",
-    "ExitCurrentPath",
     "Effect",
-    "ExternalLaunchCompleted",
-    "ExternalLaunchFailed",
-    "FileSearchCompleted",
-    "FileSearchFailed",
+    "EmptyTrashConfirmationState",
     "FileSearchResultState",
+    "FilterState",
     "FindReplaceFieldId",
     "GrepReplaceFieldId",
     "GrepReplaceSelectedFieldId",
     "GrepSearchFieldId",
-    "GrepSearchCompleted",
-    "GrepSearchFailed",
     "GrepSearchResultState",
-    "ReplaceFieldId",
-    "ReplacePreviewResultState",
-    "FileMutationCompleted",
-    "FileMutationFailed",
-    "FocusSplitTerminal",
-    "FilterState",
-    "GoBack",
-    "GoForward",
-    "GoToHomeDirectory",
-    "GoToParentDirectory",
     "HistoryState",
-    "InitializeState",
-    "MoveConfigEditorCursor",
-    "CloseSplitTerminalEffect",
     "LoadBrowserSnapshotEffect",
     "LoadChildPaneSnapshotEffect",
-    "MoveCommandPaletteCursor",
-    "JumpCursor",
-    "MoveCursor",
-    "MoveCursorAndSelectRange",
-    "MoveCursorByPage",
-    "MovePendingInputCursor",
-    "OpenNewTab",
-    "AddBookmark",
-    "OpenGrepResultInEditor",
-    "OpenFindResultInEditor",
-    "OpenPathInEditor",
-    "OpenPathWithDefaultApp",
-    "OpenTerminalAtPath",
-    "ShowAttributes",
     "NameConflictState",
     "NotificationState",
     "PaneState",
+    "PasteConflictState",
     "PendingInputState",
     "PendingKeySequenceState",
-    "ShellCommandCompleted",
-    "ShellCommandFailed",
-    "ShellCommandState",
-    "PasteClipboard",
-    "PasteConflictState",
     "ReduceResult",
-    "ReloadDirectory",
-    "RequestBrowserSnapshot",
-    "RequestDirectorySizes",
-    "RemoveBookmark",
-    "ResolvePasteConflict",
+    "ReplaceFieldId",
+    "ReplacePreviewResultState",
     "RunArchiveExtractEffect",
     "RunArchivePreparationEffect",
     "RunClipboardPasteEffect",
     "RunConfigSaveEffect",
     "RunDirectorySizeEffect",
     "RunExternalLaunchEffect",
+    "RunFileMutationEffect",
     "RunFileSearchEffect",
     "RunGrepSearchEffect",
-    "RunFileMutationEffect",
     "RunShellCommandEffect",
     "RunTextReplaceApplyEffect",
     "RunTextReplacePreviewEffect",
     "RunUndoEffect",
     "RunZipCompressEffect",
     "RunZipCompressPreparationEffect",
-    "SendSplitTerminalInput",
-    "SelectAllVisibleEntries",
-    "SetCommandPaletteQuery",
-    "SetCursorPath",
-    "SetFilterQuery",
-    "SetGrepSearchField",
-    "SetGrepReplaceField",
-    "SetGrepReplaceSelectedField",
-    "SetFindReplaceField",
-    "SetReplaceField",
-    "SetNotification",
-    "SetPendingInputCursor",
-    "SetPendingInputValue",
-    "SetPendingKeySequence",
-    "SetShellCommandValue",
-    "SetSort",
-    "SetTerminalHeight",
-    "PasteIntoPendingInput",
-    "SplitTerminalExited",
-    "SplitTerminalOutputReceived",
+    "ShellCommandState",
+    "SortState",
     "SplitTerminalFocusTarget",
-    "SplitTerminalStartFailed",
-    "SplitTerminalStarted",
     "SplitTerminalState",
     "SplitTerminalStatus",
     "StartSplitTerminalEffect",
-    "SetUiMode",
-    "SaveConfigEditor",
-    "SortState",
-    "SubmitCommandPalette",
-    "SubmitPendingInput",
-    "SubmitShellCommand",
-    "TextReplaceApplied",
-    "TextReplaceApplyFailed",
-    "TextReplacePreviewCompleted",
-    "TextReplacePreviewFailed",
-    "ToggleHiddenFiles",
-    "ToggleSplitTerminal",
-    "ToggleSelection",
-    "ToggleSelectionAndAdvance",
-    "UndoCompleted",
-    "UndoEntry",
-    "UndoFailed",
-    "UndoLastOperation",
     "UiMode",
-    "ZipCompressCompleted",
-    "ZipCompressConfirmationState",
-    "ZipCompressFailed",
-    "ZipCompressPreparationCompleted",
-    "ZipCompressPreparationFailed",
-    "ZipCompressProgress",
-    "ZipCompressProgressState",
+    "UndoEntry",
     "WriteSplitTerminalInputEffect",
-    "CycleConfigEditorValue",
-    "select_config_dialog_state",
+    "ZipCompressConfirmationState",
+    "ZipCompressProgressState",
     "build_initial_app_state",
     "build_placeholder_app_state",
     "dispatch_key_input",
     "iter_bound_keys",
     "load_browser_tab",
     "reduce_app_state",
-    "select_browser_tabs",
     "select_attribute_dialog_state",
-    "select_tab_bar_state",
+    "select_browser_tabs",
     "select_child_entries",
     "select_command_palette_state",
+    "select_config_dialog_state",
     "select_conflict_dialog_state",
-    "select_current_summary_state",
     "select_current_entries",
+    "select_current_summary_state",
     "select_help_bar_state",
     "select_input_bar_state",
     "select_input_dialog_state",
@@ -488,6 +182,7 @@ __all__ = [
     "select_shell_data",
     "select_split_terminal_state",
     "select_status_bar_state",
+    "select_tab_bar_state",
     "select_target_paths",
     "select_visible_current_entry_states",
 ]
