+++ tests/test_app_runtime.py
@@ -32,14 +32,7 @@ from zivo.models import AppConfig, ExternalLaunchRequest, UndoDeletePathStep, Un
 from zivo.services import InvalidFileSearchQueryError
 from zivo.state import (
     BrowserSnapshot,
-    BrowserSnapshotLoaded,
-    ChildPaneSnapshotLoaded,
-    ConfigSaveCompleted,
     DirectoryEntryState,
-    DirectorySizesLoaded,
-    ExternalLaunchFailed,
-    FileSearchCompleted,
-    FileSearchFailed,
     LoadBrowserSnapshotEffect,
     LoadChildPaneSnapshotEffect,
     PaneState,
@@ -49,11 +42,20 @@ from zivo.state import (
     RunFileSearchEffect,
     RunGrepSearchEffect,
     RunUndoEffect,
-    SplitTerminalStartFailed,
     StartSplitTerminalEffect,
-    UndoCompleted,
     WriteSplitTerminalInputEffect,
     build_initial_app_state,
+)
+from zivo.state.actions import (
+    BrowserSnapshotLoaded,
+    ChildPaneSnapshotLoaded,
+    ConfigSaveCompleted,
+    DirectorySizesLoaded,
+    ExternalLaunchFailed,
+    FileSearchCompleted,
+    FileSearchFailed,
+    SplitTerminalStartFailed,
+    UndoCompleted,
 )
