+++ tests/test_state_reducer.py
@@ -8,12 +8,36 @@ from zivo.models import (
     ExternalLaunchRequest,
 )
 from zivo.state import (
+    BrowserSnapshot,
+    CloseSplitTerminalEffect,
+    ConfigEditorState,
+    CurrentPaneDeltaState,
+    DirectoryEntryState,
+    DirectorySizeCacheEntry,
+    DirectorySizeDeltaState,
+    HistoryState,
+    LoadBrowserSnapshotEffect,
+    LoadChildPaneSnapshotEffect,
+    NameConflictState,
+    NotificationState,
+    PaneState,
+    PendingInputState,
+    PendingKeySequenceState,
+    RunConfigSaveEffect,
+    RunDirectorySizeEffect,
+    RunExternalLaunchEffect,
+    StartSplitTerminalEffect,
+    WriteSplitTerminalInputEffect,
+    build_initial_app_state,
+    reduce_app_state,
+    select_browser_tabs,
+)
+from zivo.state.actions import (
     ActivateNextTab,
     ActivatePreviousTab,
     AddBookmark,
     BeginFilterInput,
     BeginHistorySearch,
-    BrowserSnapshot,
     BrowserSnapshotFailed,
     BrowserSnapshotLoaded,
     CancelFilterInput,
@@ -21,18 +45,12 @@ from zivo.state import (
     ChildPaneSnapshotLoaded,
     ClearSelection,
     CloseCurrentTab,
-    CloseSplitTerminalEffect,
-    ConfigEditorState,
     ConfigSaveCompleted,
     ConfigSaveFailed,
     ConfirmFilterInput,
     CopyPathsToClipboard,
-    CurrentPaneDeltaState,
     CutTargets,
     CycleConfigEditorValue,
-    DirectoryEntryState,
-    DirectorySizeCacheEntry,
-    DirectorySizeDeltaState,
     DirectorySizesFailed,
     DirectorySizesLoaded,
     DismissConfigEditor,
@@ -45,30 +63,19 @@ from zivo.state import (
     GoForward,
     GoToHomeDirectory,
     GoToParentDirectory,
-    HistoryState,
     JumpCursor,
-    LoadBrowserSnapshotEffect,
-    LoadChildPaneSnapshotEffect,
     MoveConfigEditorCursor,
     MoveCursor,
     MoveCursorAndSelectRange,
     MoveCursorByPage,
-    NameConflictState,
-    NotificationState,
     OpenNewTab,
     OpenPathInEditor,
     OpenPathWithDefaultApp,
     OpenTerminalAtPath,
-    PaneState,
-    PendingInputState,
-    PendingKeySequenceState,
     ReloadDirectory,
     RemoveBookmark,
     RequestBrowserSnapshot,
     RequestDirectorySizes,
-    RunConfigSaveEffect,
-    RunDirectorySizeEffect,
-    RunExternalLaunchEffect,
     SaveConfigEditor,
     SendSplitTerminalInput,
     SetCursorPath,
@@ -81,14 +88,9 @@ from zivo.state import (
     SplitTerminalExited,
     SplitTerminalOutputReceived,
     SplitTerminalStarted,
-    StartSplitTerminalEffect,
     ToggleHiddenFiles,
     ToggleSelection,
     ToggleSplitTerminal,
-    WriteSplitTerminalInputEffect,
-    build_initial_app_state,
-    reduce_app_state,
-    select_browser_tabs,
 )
