+++ tests/test_state_reducer_palette_commands.py
@@ -9,32 +9,34 @@ from zivo.models import (
 )
 from zivo.state import (
     AttributeInspectionState,
-    BeginBookmarkSearch,
-    BeginCommandPalette,
-    BeginGoToPath,
-    BeginHistorySearch,
-    CancelCommandPalette,
     CommandPaletteState,
     ConfigEditorState,
     DirectoryEntryState,
-    DismissAttributeDialog,
     HistoryState,
     LoadBrowserSnapshotEffect,
-    MoveCommandPaletteCursor,
     NotificationState,
     PaneState,
     PendingInputState,
     PendingKeySequenceState,
     RunConfigSaveEffect,
     RunExternalLaunchEffect,
-    SetCommandPaletteQuery,
-    SetCursorPath,
-    ShowAttributes,
     StartSplitTerminalEffect,
-    SubmitCommandPalette,
     build_initial_app_state,
     reduce_app_state,
     select_command_palette_state,
+)
+from zivo.state.actions import (
+    BeginBookmarkSearch,
+    BeginCommandPalette,
+    BeginGoToPath,
+    BeginHistorySearch,
+    CancelCommandPalette,
+    DismissAttributeDialog,
+    MoveCommandPaletteCursor,
+    SetCommandPaletteQuery,
+    SetCursorPath,
+    ShowAttributes,
+    SubmitCommandPalette,
 )
