+++ tests/test_state_reducer_palette_replace.py
@@ -8,6 +8,21 @@ from zivo.models import (
     TextReplaceResult,
 )
 from zivo.state import (
+    DirectoryEntryState,
+    FileSearchResultState,
+    GrepSearchResultState,
+    LoadBrowserSnapshotEffect,
+    NotificationState,
+    PaneState,
+    ReplacePreviewResultState,
+    RunFileSearchEffect,
+    RunGrepSearchEffect,
+    RunTextReplaceApplyEffect,
+    RunTextReplacePreviewEffect,
+    build_initial_app_state,
+    reduce_app_state,
+)
+from zivo.state.actions import (
     BeginCommandPalette,
     BeginFindAndReplace,
     BeginGrepReplace,
@@ -16,20 +31,9 @@ from zivo.state import (
     CancelCommandPalette,
     CycleFindReplaceField,
     CycleReplaceField,
-    DirectoryEntryState,
     FileSearchCompleted,
-    FileSearchResultState,
     GrepSearchCompleted,
-    GrepSearchResultState,
-    LoadBrowserSnapshotEffect,
     MoveCommandPaletteCursor,
-    NotificationState,
-    PaneState,
-    ReplacePreviewResultState,
-    RunFileSearchEffect,
-    RunGrepSearchEffect,
-    RunTextReplaceApplyEffect,
-    RunTextReplacePreviewEffect,
     SetCommandPaletteQuery,
     SetFindReplaceField,
     SetGrepReplaceField,
@@ -40,8 +44,6 @@ from zivo.state import (
     TextReplaceApplyFailed,
     TextReplacePreviewCompleted,
     TextReplacePreviewFailed,
-    build_initial_app_state,
-    reduce_app_state,
 )
 from zivo.state.reducer_common import browser_snapshot_invalidation_paths
 
@@ -1217,4 +1219,3 @@ def test_grs_grep_search_completed_filters_non_target_results() -> None:
     )
     # Preview is triggered even with empty replace text to show find matches
     assert result.state.pending_replace_preview_request_id is not None
-
