+++ tests/test_state_reducer_palette_search.py
@@ -5,36 +5,38 @@ from zivo.models import (
     ExternalLaunchRequest,
 )
 from zivo.state import (
-    BeginCommandPalette,
-    BeginFileSearch,
-    BeginGrepSearch,
-    BeginSelectedFilesGrep,
-    CancelCommandPalette,
     CommandPaletteState,
-    CycleSelectedFilesGrepField,
     DirectoryEntryState,
-    FileSearchCompleted,
-    FileSearchFailed,
     FileSearchResultState,
-    GrepSearchCompleted,
-    GrepSearchFailed,
     GrepSearchResultState,
     LoadBrowserSnapshotEffect,
     LoadChildPaneSnapshotEffect,
     NotificationState,
-    OpenFindResultInEditor,
-    OpenGrepResultInEditor,
     PaneState,
     RunDirectorySizeEffect,
     RunExternalLaunchEffect,
     RunFileSearchEffect,
     RunGrepSearchEffect,
+    build_initial_app_state,
+    reduce_app_state,
+)
+from zivo.state.actions import (
+    BeginCommandPalette,
+    BeginFileSearch,
+    BeginGrepSearch,
+    BeginSelectedFilesGrep,
+    CancelCommandPalette,
+    CycleSelectedFilesGrepField,
+    FileSearchCompleted,
+    FileSearchFailed,
+    GrepSearchCompleted,
+    GrepSearchFailed,
+    OpenFindResultInEditor,
+    OpenGrepResultInEditor,
     SelectedFilesGrepKeywordChanged,
     SetCommandPaletteQuery,
     SetGrepSearchField,
     SubmitCommandPalette,
-    build_initial_app_state,
-    reduce_app_state,
 )
